+++ src/ast/parse.rs
@@ -8,13 +8,14 @@ use ordered_float::OrderedFloat;
 
 #[macro_export]
 macro_rules! span {
-    () => {
+    () => {{
+        use $crate::ast::{RustSpan, Span};
         Span::Rust(std::sync::Arc::new(RustSpan {
             file: file!(),
             line: line!(),
             column: column!(),
         }))
-    };
+    }};
 }
 
 // We do an unidiomatic thing here by using a struct instead of an enum.
@@ -233,7 +234,6 @@ impl Parser {
             || self.exprs.contains_key(&name)
             || self.commands.contains_key(&name)
         {
-            use egglog_ast::span::{RustSpan, Span};
             return Err(Error::CommandAlreadyExists(name, span!()));
         }
         self.user_defined.insert(name);