+++ tests/test_command_macros.rs
@@ -2,7 +2,6 @@ use crate::util::INTERNAL_SYMBOL_PREFIX;
 use egglog::ast::*;
 use egglog::util::FreshGen;
 use egglog::*;
-use egglog_ast::span::{RustSpan, Span};
 use std::sync::Arc;
 
 // Macro that prefixes rule names with a specific prefix